requirements.txt
python-dotenv==1.0.0
python-telegram-bot>=21.0
openai>=1.0.0
firebase-admin==6.3.0
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
pdfminer.six==20221105
nltk==3.8.1
numpy>=1.26.0
pandas>=2.1.0
flask==3.0.0
functions-framework==3.*